New version contains the following modifications:
logging.basicConfig is used to configure the logging level and format of log messages. A LOG_COLORS dictionary is also defined to associate colors with different logging levels.

uses a more modular structure with asynchronous functions defined for different tasks, such as crawl, get_sensitive_directories, and check_external_resources.

uses special colors and formatting in registration messages, such as using colors for different registration levels and emojis for initial impressions.

has been updated to include a function called export_to_json, which takes a data dictionary and a file name as input, and exports the data to the specified JSON file in the same directory where the script is run. Then in the scrape_data function, after extracting the relevant information, the data is exported to a JSON file using this function.

The access_ipinfo_api function is now responsible for accessing the IPinfo API and obtaining information about the given IP address or domain name.*Under review*